Fix inter hunk formatting.
Previously, only the first inter-hunk section was correcly formatted:

```
index aac5ae5..c8517bb 100644
--- src/main.rs
+++ src/main.rs
@@ -13,20 +13,11 @@ fn main() -> io::Result<()> {
```

should not highlight --- and +++ but output these without change.

So we put text attributes only after seeing a line starting with '@'
until a line that is not part of the hunk (ie, does not start with
'+', '-' or ' ').