Tokenization: skip leading whitespaces.
This change improves the presentation; spaces following the leading
'+', '-' or ' ' in junks is never colorized. It also makes the diff
computation slightly faster, as less tokens need to be processed.

* diffr-lib (TokenKind::NewLine): new case.
(tokenize, classify_byte): handle specifically spaces following the first character.

* diffr (paint_line): skip spaces a second time for coloring.
Unfortunately the info at that point does not allow to do the
processing only once.