main.rs: improve coloring contrast of white letters
The ANSI white is actually gray on many implementations. The actual
white that seem to work on all implementations is "bright white".
`termcolor` crate has no enum member for it, so we create it with Rgb().

Note that the crate has no support for Rgb() or Ansi256() functions
while used in Windows console. Documentation says:

> If they are used on Windows, then they are silently ignored and no
> colors will be emitted

This should more or less work for Windows, because before commit

    7ae151697 "main.rs: use White as a foreground for Red and Green"

…we were not setting the color foreground at all, so in that case
behavior is back to the old one. The discussion on the issue #31 being
fixed by 7ae151697 mentions that the default White behaves same way
there as it did without color being set. Thus, the behavior before and
after the current commit should be the same on Windows.

Fixes: https://github.com/mookid/diffr/issues/72