cli_args.rs: simplify arguments parser
There is no reason for config-handling hooks (like color(), html(), etc)
to peek at the value that caused them to be called. So call next()
before calling the hooks.

That allows to simplify the code and remove some args, but we do not do
that per mookid's review, who asked to leave arguments of the hooks to
be the same even if some of the functions do not actually use arguments.